requirements.txt
asgiref==3.7.2
click==8.1.6
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
